clip_l2/requirements.txt
numpy
opencv-python
torch
requests
fastapi
pydantic
uvicorn
clip-by-openai
